Move counties out of constants
